Worked on improving the code
