+++ packages/nodes-base/nodes/Aws/IAM/descriptions/UserDescription.ts
@@ -200,7 +200,7 @@ const createFields: INodeProperties[] = [
 		options: [
 			{
 				displayName: 'Path',
-				name: 'Path',
+				name: 'path',
 				type: 'string',
 				validateType: 'string',
 				default: '',
@@ -209,7 +209,7 @@ const createFields: INodeProperties[] = [
 			},
 			{
 				displayName: 'Permissions Boundary',
-				name: 'PermissionsBoundary',
+				name: 'permissionsBoundary',
 				default: '',
 				description:
 					'The ARN of the managed policy that is used to set the permissions boundary for the user',
@@ -219,7 +219,7 @@ const createFields: INodeProperties[] = [
 			},
 			{
 				displayName: 'Tags',
-				name: 'Tags',
+				name: 'tags',
 				type: 'fixedCollection',
 				description: 'A list of tags that you want to attach to the new user',
 				default: [],
@@ -257,7 +257,7 @@ const createFields: INodeProperties[] = [
 const getFields: INodeProperties[] = [
 	{
 		displayName: 'User',
-		name: 'UserName',
+		name: 'userName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
@@ -348,7 +348,7 @@ const getAllFields: INodeProperties[] = [
 		options: [
 			{
 				displayName: 'Path Prefix',
-				name: 'PathPrefix',
+				name: 'pathPrefix',
 				type: 'string',
 				validateType: 'string',
 				default: '/',
@@ -362,7 +362,7 @@ const getAllFields: INodeProperties[] = [
 const deleteFields: INodeProperties[] = [
 	{
 		displayName: 'User',
-		name: 'UserName',
+		name: 'userName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
@@ -409,7 +409,7 @@ const deleteFields: INodeProperties[] = [
 const updateFields: INodeProperties[] = [
 	{
 		displayName: 'User',
-		name: 'UserName',
+		name: 'userName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
@@ -453,7 +453,7 @@ const updateFields: INodeProperties[] = [
 	},
 	{
 		displayName: 'New User Name',
-		name: 'NewUserName',
+		name: 'newUserName',
 		default: '',
 		placeholder: 'e.g. JohnSmith',
 		type: 'string',
@@ -484,7 +484,7 @@ const updateFields: INodeProperties[] = [
 		options: [
 			{
 				displayName: 'New Path',
-				name: 'NewPath',
+				name: 'newPath',
 				type: 'string',
 				validateType: 'string',
 				default: '/',
@@ -497,7 +497,7 @@ const updateFields: INodeProperties[] = [
 const addToGroupFields: INodeProperties[] = [
 	{
 		displayName: 'User',
-		name: 'UserName',
+		name: 'userName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
@@ -541,7 +541,7 @@ const addToGroupFields: INodeProperties[] = [
 	},
 	{
 		displayName: 'Group',
-		name: 'GroupName',
+		name: 'groupName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
@@ -588,7 +588,7 @@ const addToGroupFields: INodeProperties[] = [
 const removeFromGroupFields: INodeProperties[] = [
 	{
 		displayName: 'User',
-		name: 'UserName',
+		name: 'userName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
@@ -632,7 +632,7 @@ const removeFromGroupFields: INodeProperties[] = [
 	},
 	{
 		displayName: 'Group',
-		name: 'GroupName',
+		name: 'groupName',
 		required: true,
 		type: 'resourceLocator',
 		default: {
