+++ packages/nodes-base/nodes/Aws/IAM/test/PresendStringifyBody.test.ts
@@ -1,4 +1,5 @@
-import { presendStringifyBody } from '../GenericFunctions';
+import { presendStringifyBody } from '../generalFunctions/dataHandling';
+
 describe('presendStringifyBody', () => {
 	let mockContext: any;
