+++ packages/nodes-base/nodes/Aws/IAM/test/ProcessUsersResponse.test.ts
@@ -1,4 +1,4 @@
-import { processUsersResponse } from '../GenericFunctions';
+import { processUsersResponse } from '../generalFunctions/dataHandling';
 
 describe('processUsersResponse', () => {
 	let mockContext: any;
@@ -30,7 +30,7 @@ describe('processUsersResponse', () => {
 
 		const result = await processUsersResponse.call(mockContext, [], response);
 
-		expect(result).toEqual([{ UserName: 'user1', UserId: '1' }]);
+		expect(result).toEqual([{ json: { UserName: 'user1', UserId: '1' } }]);
 	});
 
 	it('should process multiple users response correctly (ListUsersResponse)', async () => {
@@ -56,7 +56,47 @@ describe('processUsersResponse', () => {
 		const result = await processUsersResponse.call(mockContext, [], response);
 
 		expect(result).toHaveLength(2);
-		expect(result[0]).toEqual({ UserName: 'user1', UserId: '1' });
-		expect(result[1]).toEqual({ UserName: 'user2', UserId: '2' });
+		expect(result[0]).toEqual({ json: { UserName: 'user1', UserId: '1' } });
+		expect(result[1]).toEqual({ json: { UserName: 'user2', UserId: '2' } });
+	});
+
+	it('should return an empty array if no users found in GetUserResponse', async () => {
+		const response = {
+			body: {
+				GetUserResponse: {
+					GetUserResult: {},
+					ResponseMetadata: {},
+				},
+			},
+			statusCode: 200,
+			statusMessage: 'OK',
+			headers: {},
+		};
+
+		mockGetNodeParameter.mockReturnValue('get');
+
+		const result = await processUsersResponse.call(mockContext, [], response);
+
+		expect(result).toEqual([]);
+	});
+
+	it('should return an empty array if no users found in ListUsersResponse', async () => {
+		const response = {
+			body: {
+				ListUsersResponse: {
+					ListUsersResult: { Users: [] },
+					ResponseMetadata: {},
+				},
+			},
+			statusCode: 200,
+			statusMessage: 'OK',
+			headers: {},
+		};
+
+		mockGetNodeParameter.mockReturnValue('getAll');
+
+		const result = await processUsersResponse.call(mockContext, [], response);
+
+		expect(result).toEqual([]);
 	});
 });
